--- src/main/java/com/spring/ideafestivalbackend/GHC/domain/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.spring.ideafestivalbackend.GHC.domain.user.repository;
-import com.spring.ideafestivalbackend.GHC.domain.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
