+++ src/main/java/com/spring/ideafestivalbackend/domain/board/entity/BoardEntity.java
@@ -10,6 +10,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+@Entity
 @Getter
 @Builder
 @AllArgsConstructor
