--- src/main/java/com/spring/ideafestivalbackend/domain/board/exception/handler/GlobalExceptionHandler.java
@@ -1,24 +1,0 @@
-package com.spring.ideafestivalbackend.domain.board.exception.handler;
-
-import com.spring.ideafestivalbackend.domain.board.exception.ErrorCode;
-import com.spring.ideafestivalbackend.domain.board.exception.ErrorMessage;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import javax.servlet.http.HttpServletRequest;
-
-@ControllerAdvice
-@Slf4j
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorMessage> methodValidException(MethodArgumentNotValidException e, HttpServletRequest request){
-        log.warn("MethodArgumentNotValidException 발생!!! url:{}, trace:{}",request.getRequestURI(), e.getStackTrace());
-        ErrorMessage errorMessage = new ErrorMessage(ErrorCode.NOT_NULL.getMessage(), ErrorCode.NOT_NULL.getStatus());   //makeErrorResponse(e.getBindingResult());
-        return new ResponseEntity<ErrorMessage>(errorMessage, HttpStatus.BAD_REQUEST);
-    }
-}
