+++ src/main/java/com/spring/ideafestivalbackend/domain/board/repository/BoardRepository.java
@@ -1,4 +1,4 @@
-package com.spring.ideafestivalbackend.domain.board.exception.repository;
+package com.spring.ideafestivalbackend.domain.board.repository;
