+++ src/main/java/com/spring/ideafestivalbackend/domain/user/exception/UserNotFoundException.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class UserNotFoundException extends RuntimeException{
     private final ErrorCode errorCode;
-
+    //
     public UserNotFoundException(String message){
         super(message);
         this.errorCode = ErrorCode.USER_NOT_FOUD;
