+++ src/main/resources/application.yml
@@ -22,8 +22,8 @@ spring:
           auth: 'true'
     password: gree7942
   redis:
-    host: 127.0.0.1
-    port: '6379'
+    host: localhost
+    port: 6379
   main:
     allow-bean-definition-overriding: 'true'
